test(gateway): isolate terminal task lifecycle coverage
